crypto_arbitrage_bot_ready/requirements.txt
python-telegram-bot==20.8
aiohttp
python-dotenv
openai
ccxt
uvloop